frontend/public/AGENT/pyproject.toml
[project]
name = "agenteui"
version = "0.1.0"
description = "Sistema de agente IA para barbería"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "mcp>=1.0.0",
    "mirascope[google]>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.1.1",
    "asyncpg>=0.30.0",
]